Add classes and methods of: Person, Student & Teacher
