+++ person_class.rb
@@ -17,25 +17,3 @@ class Person
     end
   end
 end
-
-class Student < Person
-  def
-    super(_name, _age, _parent_permission)
-    @classroom
-  end
-
-  def play_hooky
-    '¯\(ツ)/¯'
-  end
-end
-
-class Teacher < Person
-  def
-    super(_name, _age)
-    @specialization
-  end
-
-  def can_use_services?
-    true
-  end
-end
