+++ student_class.rb
@@ -1,0 +1,12 @@
+require_relative 'person_class'
+
+class Student < Person
+  def
+    super(_name, _age, _parent_permission)
+    @classroom
+  end
+
+  def play_hooky
+    '¯\(ツ)/¯'
+  end
+end
