+++ teacher_class.rb
@@ -1,0 +1,12 @@
+require_relative 'person_class'
+
+class Teacher < Person
+  def
+    super(_name, _age)
+    @specialization
+  end
+
+  def can_use_services?
+    true
+  end
+end
